PostView2: Fixed plotmix tool.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/PostView2@10913 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ PostView2/GLPlaneCutPlot.cpp
@@ -937,7 +937,7 @@ void CGLPlaneCutPlot::ReleasePlane()
 
 int CGLPlaneCutPlot::ClipPlanes()
 {
-	return m_pcp.size();
+	return (int)m_pcp.size();
 }
 
 CGLPlaneCutPlot* CGLPlaneCutPlot::GetClipPlane(int i)
